refactor: move cosineSimilarity function to clientHelpers
- Relocated the cosineSimilarity function from utils.ts to a new clientHelpers module to improve code organization and modularity.
